Project MyStudy
This is a web2 - elearning project on based online teaching platform, which has 3 types of users: admin, teachers & students. An admin can edit permissions, access and all in the website, a teacher can upload quizes, upload notes, pdf and all for students, a student can appear quizes, download study materials and watch lecture videos.

This is a under develop partially made project, which can be further improved to have more features, easy to use and have access to different things.

Co-Authored-By: Elias Hernandis <2157787+https@users.noreply.github.com>